Add Player creating
